Исправлен баг, приводящий к возникновению бесконечного цикла из-за некорректного определения вхождения дочернего узла в родительский в результате декомпрессии. Исправлены баги с некорректным обнаружением общих родительских узлов квадродеревьев разных уровней в результате декомпрессии. Рефакторинг.

+++ Assets/SpatialPartitionSystem/Core/Scripts/Series/Node.cs
@@ -3,10 +3,10 @@
     internal struct Node
     {
         public int firstChildIndex;
-        public ushort objectsCount;
+        public byte objectsCount;
         public bool isLeaf;
         
-        public sbyte depth;
+        public byte depth;
         public AABB2D bounds;
     }
 }
